requirements.txt
discord-py-interactions==5.13.2
discord-typings==0.9.0
discord.py==2.5.2
Flask==3.1.0
psycopg2-binary
